hotfix: fix commonLinks visible on mobile
