+++ components/CommonLinks/CommonLinks.jsx
@@ -5,10 +5,11 @@ import rgbaToRgb from 'rgba-to-rgb'
 import classNames from 'classnames'
 
 import styles from './CommonLinks.module.css'
-import { ThemeContext } from '../../pages/_app'
+import { ThemeContext, OpenTocContext } from '../../pages/_app'
 
 export function CommonLinks({ color, bgColor, mainPage }) {
     const isDark = useContext(ThemeContext)
+    const { isOpen } = useContext(OpenTocContext)
     const mainPageColor = isDark ? '#1A1C1F' : '#f5f8fb'
     const commonColor = mainPage
         ? mainPageColor
@@ -20,6 +21,7 @@ export function CommonLinks({ color, bgColor, mainPage }) {
           )
     const containerClassnames = classNames(styles.CommonLinks__container, {
         [styles.mainPageContainer]: mainPage,
+        [styles.hidden]: !isOpen,
     })
 
     return (
