+++ components/HamburgerMenu/menu.module.css
@@ -16,6 +16,7 @@
 .line {
     fill: none;
     stroke-width: 6;
+    stroke: black;
     transition: stroke-dasharray 600ms cubic-bezier(0.4, 0, 0.2, 1),
         stroke-dashoffset 600ms cubic-bezier(0.4, 0, 0.2, 1);
 }
@@ -23,34 +24,40 @@
 .line1 {
     stroke-dasharray: 60 207;
     stroke-width: 6;
+    stroke: black;
 }
 
 .line2 {
     stroke-dasharray: 60 60;
     stroke-width: 6;
+    stroke: black;
 }
 
 .line3 {
     stroke-dasharray: 60 207;
     stroke-width: 6;
+    stroke: black;
 }
 
 .opened .line1 {
     stroke-dasharray: 90 207;
     stroke-dashoffset: -134;
     stroke-width: 6;
+    stroke: black;
 }
 
 .opened .line2 {
     stroke-dasharray: 1 60;
     stroke-dashoffset: -30;
     stroke-width: 6;
+    stroke: black;
 }
 
 .opened .line3 {
     stroke-dasharray: 90 207;
     stroke-dashoffset: -134;
     stroke-width: 6;
+    stroke: black;
 }
 
 @media print {
